New method to fix the "no kernel" issue

+++ scripts/remaster_rootfs.sh
@@ -69,10 +69,26 @@ function install_packages()
 	
 	#Some daily images do not have a kernel;
 	#ensure that a valid kernel is installed
-	KERNEL_PKG=linux-signed-generic-lts-trusty
-	[ "$(uname -m)" == "x86_64" ] || KERNEL_PKG=linux-image-generic-lts-trusty
-	apt-get --reinstall -y install $KERNEL_PKG
-	apt-cache depends $KERNEL_PKG | tail -n+2 | awk '{print $NF}' | xargs apt-get --reinstall -y install
+
+	#make sure we have a initrd (otherwise the kernel update may fail
+	if [ ! -e "$(readlink -f /initrd.img)" ]; then
+		echo "LIPCK: No initrd in place; generating new one."
+		update-initramfs -v -c -k all
+	fi
+
+	#Note: this does only work if we have a recent install iso since older kernels are removed from
+	# the repositories
+	KERNEL_PKG=$(dpkg -S "$(readlink -f /vmlinuz)" | cut -d ":" -f1)
+	if [ -z "$KERNEL_PKG" ]; then
+		echo "LIPCK: remaster_rootfs: unable to determine installed kernel version; giving up..."
+	fi
+	#[ "$(uname -m)" == "x86_64" ] || KERNEL_PKG=linux-image-generic-lts-trusty
+	if [ ! -e "$(readlink -f /initrd.img)" ]; then
+                echo "LIPCK: No kernel in place; try to reinstall kernel image package:"
+		echo "       $KERNEL_PKG"
+		apt-get --reinstall -y install $KERNEL_PKG
+		#apt-cache depends $KERNEL_PKG | tail -n+2 | awk '{print $NF}' | xargs apt-get --reinstall -y install
+	fi
 
 	install_packages_from_file "$CONTRIB_DIR/pre_installed_packages" ""
 	install_packages_from_file "$CONTRIB_DIR/pre_installed_packages.without-recommends" "--no-install-recommends"
